[refactor] : application.common 안의 dto와 mapper -> surevy 하위로 변경
