+++ survey/application/src/main/java/me/nalab/survey/application/common/survey/dto/FormQuestionDtoable.java
@@ -1,4 +1,4 @@
-package me.nalab.survey.application.common.dto;
+package me.nalab.survey.application.common.survey.dto;
 
 import java.time.LocalDateTime;
