+++ survey/application/src/main/java/me/nalab/survey/application/port/in/web/CreateSurveyUseCase.java
@@ -1,6 +1,6 @@
 package me.nalab.survey.application.port.in.web;
 
-import me.nalab.survey.application.common.dto.SurveyDto;
+import me.nalab.survey.application.common.survey.dto.SurveyDto;
 
 /**
  * 새로운 Survey를 생성하는 Usecase 이 인터페이스를 이용해서 새로운 Survey를 생성할 수 있음.
