+++ survey/application/src/test/java/me/nalab/survey/application/RandomSurveyDtoFixture.java
@@ -7,14 +7,14 @@ import java.util.Random;
 import java.util.function.Supplier;
 
 import lombok.Setter;
-import me.nalab.survey.application.common.dto.ChoiceDto;
-import me.nalab.survey.application.common.dto.ChoiceFormQuestionDto;
-import me.nalab.survey.application.common.dto.ChoiceFormQuestionDtoType;
-import me.nalab.survey.application.common.dto.FormQuestionDtoable;
-import me.nalab.survey.application.common.dto.QuestionDtoType;
-import me.nalab.survey.application.common.dto.ShortFormQuestionDto;
-import me.nalab.survey.application.common.dto.ShortFormQuestionDtoType;
-import me.nalab.survey.application.common.dto.SurveyDto;
+import me.nalab.survey.application.common.survey.dto.ChoiceDto;
+import me.nalab.survey.application.common.survey.dto.ChoiceFormQuestionDto;
+import me.nalab.survey.application.common.survey.dto.ChoiceFormQuestionDtoType;
+import me.nalab.survey.application.common.survey.dto.FormQuestionDtoable;
+import me.nalab.survey.application.common.survey.dto.QuestionDtoType;
+import me.nalab.survey.application.common.survey.dto.ShortFormQuestionDto;
+import me.nalab.survey.application.common.survey.dto.ShortFormQuestionDtoType;
+import me.nalab.survey.application.common.survey.dto.SurveyDto;
 
 public class RandomSurveyDtoFixture {
