+++ survey/web-adaptor/src/main/java/me/nalab/survey/web/adaptor/createsurvey/SurveyCreateRequestMapper.java
@@ -5,14 +5,14 @@ import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import me.nalab.survey.application.common.dto.ChoiceDto;
-import me.nalab.survey.application.common.dto.ChoiceFormQuestionDto;
-import me.nalab.survey.application.common.dto.ChoiceFormQuestionDtoType;
-import me.nalab.survey.application.common.dto.FormQuestionDtoable;
-import me.nalab.survey.application.common.dto.QuestionDtoType;
-import me.nalab.survey.application.common.dto.ShortFormQuestionDto;
-import me.nalab.survey.application.common.dto.ShortFormQuestionDtoType;
-import me.nalab.survey.application.common.dto.SurveyDto;
+import me.nalab.survey.application.common.survey.dto.ChoiceDto;
+import me.nalab.survey.application.common.survey.dto.ChoiceFormQuestionDto;
+import me.nalab.survey.application.common.survey.dto.ChoiceFormQuestionDtoType;
+import me.nalab.survey.application.common.survey.dto.FormQuestionDtoable;
+import me.nalab.survey.application.common.survey.dto.QuestionDtoType;
+import me.nalab.survey.application.common.survey.dto.ShortFormQuestionDto;
+import me.nalab.survey.application.common.survey.dto.ShortFormQuestionDtoType;
+import me.nalab.survey.application.common.survey.dto.SurveyDto;
 import me.nalab.survey.web.adaptor.createsurvey.request.ChoiceFormQuestionRequest;
 import me.nalab.survey.web.adaptor.createsurvey.request.ChoiceRequest;
 import me.nalab.survey.web.adaptor.createsurvey.request.FormQuestionRequestable;
