+++ survey/web-adaptor/src/main/java/me/nalab/survey/web/adaptor/createsurvey/request/ShortFormQuestionRequest.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
-import me.nalab.survey.application.common.dto.QuestionDtoType;
+import me.nalab.survey.application.common.survey.dto.QuestionDtoType;
 
 @Getter
 @ToString
