+++ survey/web-adaptor/src/main/java/me/nalab/survey/web/adaptor/survey/find/mapper/SurveyFindResponseMapper.java
@@ -2,11 +2,11 @@ package me.nalab.survey.web.adaptor.survey.find.mapper;
 
 import java.util.stream.Collectors;
 
-import me.nalab.survey.application.common.dto.ChoiceFormQuestionDto;
-import me.nalab.survey.application.common.dto.FormQuestionDtoable;
-import me.nalab.survey.application.common.dto.ShortFormQuestionDto;
-import me.nalab.survey.application.common.dto.SurveyDto;
-import me.nalab.survey.application.common.dto.TargetDto;
+import me.nalab.survey.application.common.survey.dto.ChoiceFormQuestionDto;
+import me.nalab.survey.application.common.survey.dto.FormQuestionDtoable;
+import me.nalab.survey.application.common.survey.dto.ShortFormQuestionDto;
+import me.nalab.survey.application.common.survey.dto.SurveyDto;
+import me.nalab.survey.application.common.survey.dto.TargetDto;
 import me.nalab.survey.web.adaptor.survey.find.response.ChoiceFormQuestionResponse;
 import me.nalab.survey.web.adaptor.survey.find.response.ChoiceResponse;
 import me.nalab.survey.web.adaptor.survey.find.response.FormQuestionResponseable;
